Fix issues when deploying sls template with custom event bus * Update Event::Rule service model * Update serverless and plugin version * Update integration test
Issue fixed:
#3872 EventBridge trigger on localstack defaults to default event bus even though a new one is passed.
